+++ CHANGES.md
@@ -58,6 +58,7 @@ Release Notes.
 * Support datasource metric analysis.
 * [Break Change] Keep the endpoint avg resp time meter name the same with others scope. (This may break 3rd party integration and existing alarm rule settings)
 * Add Python FastApi component ID(7014).
+* Support all metrics from MAL engine in alarm core, including Prometheus, OC receiver, meter receiver.
 
 #### UI